Fully Compilable and Complete USB RX Module
No doubt there are plenty of bugs and issues within it, but after modifying it a lot, it is finally ready. I'm only re uploading the modules that would not compile before or under went a lot of changes. Almost anything that was fine stays the same. Also a little guide is in here for understanding step by step what should happen in the RX block and the RCU. Let me know how testing it goes. (Also please let me know if anything is missing.)